Refactor push
