+++ SharpUp/Classes/VulnerabilityCheck.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using SharpUp.Interfaces;
+
+namespace SharpUp.Classes
+{
+    public abstract class VulnerabilityCheck : IVulnerabilityCheck
+    {
+        protected bool _isVulnerable = false;
+        protected List<string> _details = new List<string>();
+        protected string _name;
+        public virtual bool IsVulnerable()
+        {
+            return _isVulnerable;
+        }
+
+        public virtual string[] Details()
+        {
+            return _details.ToArray();
+        }
+
+        public virtual string Name()
+        {
+            return _name;
+        }
+    }
+}
