+++ SharpUp/Interfaces/IVulnerabilityCheck.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace SharpUp.Interfaces
+{
+    public interface IVulnerabilityCheck
+    {
+        string Name();
+        bool IsVulnerable();
+        string[] Details();
+    }
+}
